requirements.txt
rembg==2.0.50
Pillow==10.0.0
numpy==1.24.3
